Release/1.2.7

+++ SWGoH.Model/Model/Arena.cs
@@ -7,7 +7,7 @@ namespace SWGoH.Model
     public class Arena
     {
         public int CurrentRank { get; set; }
-        public int AverageRank { get; set; }
+        public double AverageRank { get; set; }
         public List<string> ArenaTeam { get; set; }
     }
 }
